resolves warning  svn merge -r 145:146 https://amethyst.yui.ne.jp/svn-dev/x1/xmil/branches/yui/WORK_01

+++ xmil/trunk/nevent.c
@@ -62,7 +62,7 @@ void nevent_progress(void) {
 	while(wi < witem) {
 		item = *wi++;
 		if (item->proc != NULL) {
-			item->proc(item - nevent.item);
+			item->proc((UINT)(item - nevent.item));
 		}
 	}
 }
